requirements.txt.txt
python-telegram-bot==13.15
PyMuPDF